Updated report plot interactivity
