+++ neuroscout/tasks/utils/build.py
@@ -7,17 +7,11 @@ from tempfile import mkdtemp
 from bids.analysis import Analysis as BIDSAnalysis
 from bids.layout import BIDSLayout
 from copy import deepcopy
-from celery.utils.log import get_task_logger
 from grabbit.extensions.writable import build_path
-
-from ...utils.db import dump_pe
-from ...models import Analysis, PredictorEvent, Predictor, RunStimulus
-from ...schemas.analysis import AnalysisFullSchema, AnalysisResourcesSchema
 
 
 PATHS = ['sub-{subject}_[ses-{session}_]task-{task}_[acq-{acquisition}_]'
          '[run-{run}_]events.tsv']
-logger = get_task_logger(__name__)
 
 
 def writeout_events(analysis, pes, outdir):
@@ -104,77 +98,6 @@ def build_analysis(analysis, predictor_events, bids_dir, run_id=None,
         bids_analysis.setup(**entities)
 
     return tmp_dir, paths, bids_analysis
-
-
-def create_pes(predictors, run_ids):
-    """ Create PredictorEvents from EFs """
-    all_pes = []
-    for pred in predictors:
-        ef = pred.extracted_feature
-        # For all instances for stimuli in this task's runs
-        for ee in ef.extracted_events:
-            # if ee.value:
-            query = RunStimulus.query.filter_by(stimulus_id=ee.stimulus_id)
-            if run_ids is not None:
-                query = query.filter(RunStimulus.run_id.in_(run_ids))
-            for rs in query:
-                duration = ee.duration
-                if duration is None:
-                    duration = rs.duration
-                all_pes.append(
-                    dict(
-                        onset=(ee.onset or 0) + rs.onset,
-                        value=ee.value,
-                        object_id=ee.object_id,
-                        duration=duration,
-                        predictor_id=pred.id,
-                        run_id=rs.run_id,
-                        stimulus_id=ee.stimulus_id
-                    )
-                )
-    return all_pes
-
-
-def dump_analysis(analysis_id, run_id=None):
-    """" Serialize analysis and related PredictorEvents to JSON.
-    Queries PredictorEvents to get all events for all runs and predictors. """
-
-    # Query for analysis
-    analysis = Analysis.query.filter_by(hash_id=analysis_id).one()
-
-    # Dump analysis JSON
-    analysis_json = AnalysisFullSchema().dump(analysis)[0]
-    resources_json = AnalysisResourcesSchema().dump(analysis)[0]
-
-    # Get run IDs
-    all_runs = [r['id'] for r in analysis_json['runs']]
-    if run_id is None:
-        run_id = all_runs
-    if not set(run_id) <= set(all_runs):
-        raise ValueError("Incorrect run id specified")
-
-    # Query and dump PredictorEvents
-    all_pred_ids = [(p['id']) for p in analysis_json['predictors']]
-    all_preds = Predictor.query.filter(Predictor.id.in_(all_pred_ids))
-
-    base_pred_ids = [p.id for p in all_preds.filter_by(ef_id=None)]
-    ext_preds = Predictor.query.filter(
-        Predictor.id.in_(set(all_pred_ids) - set(base_pred_ids)))
-
-    pes = PredictorEvent.query.filter(
-        (PredictorEvent.predictor_id.in_(base_pred_ids)) &
-        (PredictorEvent.run_id.in_(run_id)))
-    pes = dump_pe(pes)
-
-    pes += create_pes(ext_preds, run_id)
-
-    dataset_path = Path(analysis.dataset.local_path)
-    preproc_path = dataset_path / 'derivatives' / 'fmriprep'
-
-    if preproc_path.exists():
-        dataset_path = preproc_path
-    return (analysis.id, analysis_json, resources_json, pes,
-            str(dataset_path))
 
 
 def get_entities(run):
